fix listing historial

+++ web/historial.php
@@ -25,14 +25,13 @@ $filtro = $_GET['filtro'];
 <table style="width:100%">
     <h3><tr><td>Descripcion</td><td>Link</td><td>Link Video Original</td><td>Tiempo de inicio del video original</td><td>Tiempo fin del video original</td></tr></h3>
     <?php
-    if ($filtro==NULL) {
     $result = mysqli_query($db_link, "SELECT usuario,descripcion,link,videoid,inicio,fin FROM links");
     if (mysqli_num_rows($result) > 0) {
         // output data of each row
         while($row = mysqli_fetch_assoc($result)) {
           //check_clip_existance($row["link"]);
           
-          if (check_clip_existance($row["link"]) == 1) {
+          if (check_clip_existance($row["link"]) == 1) && ($filtro == "activo") {
             $link_download="https://clipcms.unomedios.com.ar/".$row["link"];
             echo "<tr><td>".$row["descripcion"]."</td><td><a href='".$link_download. "'>Descargar clip</a></td><td><a href='https://mediacms.unomedios.com.ar/view?m=".$row["videoid"]."'>Ir al original</a></td><td>".$row["inicio"]. "</td><td>".$row["fin"]. "</td></tr>";
           } else {
@@ -45,7 +44,7 @@ $filtro = $_GET['filtro'];
       } else {
         echo "0 results";
       }
-    }
+    
     ?>
     
 </table>
